Reestrutura as files das views.Apaga files inutilizadas. Remove prints
