+++ Artero/Views/OnboardingView.swift
@@ -15,10 +15,10 @@ struct OnboardingView: View {
         let yExtension: CGFloat = 50
         GeometryReader { g in
             TabView {
-                OnBoardingScreen1()
-                OnBoardingScreen2()
-                OnBoardingScreen3()
-                OnBoardingScreen4()
+                Onboarding1View()
+                Onboarding2View()
+                Onboarding3View()
+                Onboarding4View()
                 
             }
             .frame(width: g.size.width, height: g.size.height + yExtension)
@@ -35,113 +35,5 @@ struct OnBoardingView_Previews: PreviewProvider {
     static var previews: some View {
         OnboardingView()
             .preferredColorScheme(.dark)
-    }
-}
-
-struct OnBoardingScreen1: View {
-    var body: some View {
-        ZStack {
-            Image("onboarding1")
-                .resizable()
-                .scaledToFill()
-            VStack(alignment: .leading){
-                Text(NSLocalizedString("OnBoarding1_line1", comment:""))
-                Text(NSLocalizedString("OnBoarding1_line2", comment:""))
-                Text(NSLocalizedString("OnBoarding1_line3", comment:""))
-            }.padding(.bottom, 150)
-            .padding(.leading)
-            .font(.system(size: 48, weight: .heavy, design: .default))
-            .foregroundColor(.black)
-        }
-        
-    }
-}
-struct OnBoardingScreen2: View {
-    var body: some View {
-        ZStack {
-            Image("onboarding2")
-                .resizable()
-                .scaledToFill()
-            VStack(alignment: .center){
-                Text(NSLocalizedString("OnBoarding2_line1", comment:""))
-                Text(NSLocalizedString("OnBoarding2_line2", comment:""))
-                Image(NSLocalizedString("OnBoarding2_image1", comment:""))
-                    .padding()
-                Text(NSLocalizedString("OnBoarding2_line3", comment:""))
-                Text(NSLocalizedString("OnBoarding2_line4", comment:""))
-                Text(NSLocalizedString("OnBoarding2_line5", comment:""))
-                Image(NSLocalizedString("OnBoarding2_image2", comment:""))
-                
-            }
-            .font(.system(size: 28, weight: .heavy, design: .default))
-            .foregroundColor(.black)
-        }
-    }
-}
-struct OnBoardingScreen3: View {
-    var body: some View {
-        ZStack {
-            Image("onboarding3")
-                .resizable()
-                .scaledToFill()
-            VStack(alignment: .center){
-                Image(NSLocalizedString("OnBoarding3_image1", comment: ""))
-                Image(NSLocalizedString("OnBoarding3_image2", comment: ""))
-                Text(NSLocalizedString("OnBoarding3_line1", comment:""))
-                Text(NSLocalizedString("OnBoarding3_line2", comment:""))
-            }
-            .font(.system(size: 28, weight: .heavy, design: .default))
-            .foregroundColor(.black)
-        }
-    }
-}
-struct OnBoardingScreen4: View {
-    @State private var linkActive: Bool = false
-    @StateObject private var user: User = UserController().get()
-    
-    var body: some View {
-        ZStack {
-            Image("onboarding4")
-                .resizable()
-                .scaledToFill()
-            VStack(){
-                VStack(alignment: .leading){
-                    Text(NSLocalizedString("OnBoarding4_line1", comment:""))
-                    Text(NSLocalizedString("OnBoarding4_line2", comment:""))
-                    VStack{
-                        Image(NSLocalizedString("OnBoarding4_image1", comment: ""))
-                            .padding(.trailing, 30)
-                    }
-                    Text(NSLocalizedString("OnBoarding4_line3", comment:""))
-                    Text(NSLocalizedString("OnBoarding4_line4", comment:""))
-                    Text(NSLocalizedString("OnBoarding4_line5", comment:""))
-                }
-                .padding(.top, 80)
-                .padding(.bottom, 110)
-                .padding(.trailing, 20)
-                .font(.system(size: 32, weight: .heavy, design: .default))
-                .foregroundColor(.black)
-                
-                NavigationLink(
-                    destination: HomeView(viewModel: HomeViewModel(activityRepository: UserDefaultsActivityRepository.shared, streakRepository: UserDefaultsStreakRepository.shared, themeIndexRepository: UserDefaultsThemeIndexRepository.shared)),
-                    isActive: $linkActive,
-                    label : {
-                        Text(NSLocalizedString("button_text", comment: ""))
-                            .frame(width: 310, height: 54, alignment: .center)
-                            .background(Color.white)
-                            .foregroundColor(.black)
-                            .font(.system(size: 20, weight: .semibold, design: .default))
-                            .cornerRadius(10)
-                            .shadow(color: Color("shadow"),radius: 5)
-                    }
-                )
-                .onChange(of: linkActive, perform: { value in
-                    if (linkActive) {
-                        user.onboardingComplete = true
-                        user.save()
-                    }
-                })
-            }
-        }
     }
 }
