+++ Artero/Views/ThemeView.swift
@@ -25,26 +25,26 @@ struct ThemeView: View {
         if let theme = theme {
             ZStack {
                 ScrollView {
-                    ThemeHeaderView(theme: theme)
-                    ThemeTextView(theme: theme)
+                    HeaderThemeView(theme: theme)
+                    TextThemeView(theme: theme)
                     
                 }
                 VStack {
                     Spacer()
                     if viewModel.currentDayActivity != nil {
-                        FakeButtonView()
+                        FakeButtonThemeView()
                     } else {
                         if selectedImage != nil {
-                            ConfirmButtonView(theme: theme, selectedImage: $selectedImage, isFeedbackShowing: $isFeedbackShowing) { activity in
+                            ConfirmButtonThemeView(theme: theme, selectedImage: $selectedImage, isFeedbackShowing: $isFeedbackShowing) { activity in
                                 viewModel.save(activity: activity)
                             }
                         } else {
-                            CameraButtonView(sourceType: $sourceType, isImagePickerDisplay: $isImagePickerDisplay)
+                            CameraButtonThemeView(sourceType: $sourceType, isImagePickerDisplay: $isImagePickerDisplay)
                         }
                     }
                 }
                 if isFeedbackShowing {
-                    FeedbackView(isGalleryShowing: $isGalleryShowing, isFeedbackShowing: $isFeedbackShowing)
+                    FeedbackThemeView(isGalleryShowing: $isGalleryShowing, isFeedbackShowing: $isFeedbackShowing)
                 }
                 NavigationLink(destination: GalleryView(viewModel: galleryViewModel), isActive: $isGalleryShowing, label: {})
             }
@@ -54,322 +54,5 @@ struct ThemeView: View {
                 ImagePickerView(selectedImage: self.$selectedImage, sourceType: self.$sourceType)
             }
         }
-    }
-}
-
-struct FeedbackView: View {
-    @Binding var isGalleryShowing: Bool
-    @Binding var isFeedbackShowing: Bool
-    
-    var body: some View {
-        VStack {
-            Image(systemName: "checkmark.circle.fill")
-                .font(.system(size: 80, weight: .regular, design: .default))
-            Text(NSLocalizedString("activity_saved_on", comment: ""))
-                .font(.system(size: 20, weight: .bold, design: .default))
-                .padding(.top, 5)
-                .padding(.bottom, -3)
-            Text(NSLocalizedString("gallery", comment: ""))
-                .font(.system(size: 20, weight: .bold, design: .default))
-                .foregroundColor(Color(UIColor.link))
-                .onTapGesture {
-                    isGalleryShowing = true
-                    isFeedbackShowing = false
-                }
-        }
-        .frame(width: 230, height: 230)
-        .background(Color(UIColor.systemGray6).opacity(0.9))
-        .cornerRadius(14)
-        .zIndex(1)
-    }
-}
-
-struct CameraButtonView: View {
-    @Binding var sourceType: UIImagePickerController.SourceType
-    @Binding var isImagePickerDisplay: Bool
-    @State var showEnableCameraAlert = false
-    
-    private func checkCameraPermissions() {
-        let cameraAuthorizationStatus = AVCaptureDevice.authorizationStatus(for: .video)
-        switch cameraAuthorizationStatus {
-        case .notDetermined:
-            self.askCameraPermission()
-        case .authorized:
-            self.showCamera()
-        case .restricted, .denied:
-            self.showCameraAccessAlert()
-        @unknown default:
-            self.showCameraAccessAlert()
-        }
-    }
-    
-    private func enableCameraAlert() -> Alert {
-        return Alert (
-            title: Text(NSLocalizedString("enable_camera_alert_title", comment: "")),
-            message: Text(NSLocalizedString("enable_camera_alert_description", comment: "")),
-            primaryButton: .default(Text(NSLocalizedString("settigns", comment: "")), action: {
-                UIApplication.shared.open(URL(string: UIApplication.openSettingsURLString)!)
-            }),
-            secondaryButton: .cancel()
-        )
-    }
-    
-    private func askCameraPermission() {
-        AVCaptureDevice.requestAccess(for: .video, completionHandler: { accessGranted in
-            if accessGranted {
-                self.showCamera()
-            } else {
-                self.showCameraAccessAlert()
-            }
-        })
-    }
-    
-    private func showCameraAccessAlert() {
-        self.showEnableCameraAlert = true
-    }
-    
-    private func showCamera() {
-        self.sourceType = .camera
-        self.isImagePickerDisplay.toggle()
-    }
-    
-    
-    var body: some View {
-        Menu(content: {
-            Button(action: {
-                self.checkCameraPermissions()
-            }, label: {
-                HStack {
-                    Image(systemName: "camera")
-                    Text(NSLocalizedString("button_take_a_picture", comment: ""))
-                }
-            })
-            Button(action: {
-                self.sourceType = .photoLibrary
-                self.isImagePickerDisplay.toggle()
-            }, label: {
-                HStack {
-                    Image(systemName: "photo.on.rectangle.angled")
-                    Text(NSLocalizedString("button_choose_media", comment: ""))
-                }
-            })
-        }, label: {
-            HStack {
-                Image(systemName: "camera")
-                Text(NSLocalizedString("button_media", comment: ""))
-            }
-            .frame(maxWidth: .infinity, minHeight: 50)
-            .font(.system(size: 17, weight: .semibold, design: .default))
-            .foregroundColor(.white)
-            .background(Color(.link))
-        })
-        .background(Color(UIColor.systemGray6))
-        .cornerRadius(10.0)
-        .padding(.horizontal)
-        .padding(.bottom, UIScreen.main.bounds.height * 0.05)
-        .alert(isPresented: $showEnableCameraAlert, content: { self.enableCameraAlert() })
-    }
-}
-
-struct FakeButtonView: View {
-    var body: some View {
-        Button(action: {
-            
-        }, label: {
-            HStack {
-                Text(NSLocalizedString("art_sent", comment: ""))
-            }
-            .frame(maxWidth: .infinity, minHeight: 50)
-            .font(.system(size: 17, weight: .semibold, design: .default))
-            .foregroundColor(.white)
-            .background(Color(.gray))
-            .cornerRadius(10.0)
-        })
-        .padding(.horizontal)
-        .padding(.bottom, UIScreen.main.bounds.height * 0.05)
-        .disabled(true)
-    }
-}
-
-
-struct ConfirmButtonView: View {
-    func saveActivity() {
-        guard let theme = self.theme,
-              let image = self.selectedImage
-        else {
-            return;
-        }
-        let activity = Activity(theme: theme, date: Date(), image: image)
-        onSave(activity)
-
-    }
-    
-    let theme: Theme?
-    
-    @Binding var selectedImage: UIImage?
-    @Binding var isFeedbackShowing: Bool
-    let onSave: (Activity) -> Void
-
-    var body: some View {
-        Menu(content: {
-            Button(action: {
-                selectedImage = nil
-            }, label: {
-                HStack {
-                    Image(systemName: "xmark.circle")
-                    Text(NSLocalizedString("cancel", comment: ""))
-                }
-            })
-            Button(action: {
-                isFeedbackShowing = true
-                self.saveActivity()
-                
-                let seconds = 3.0
-                DispatchQueue.main.asyncAfter(deadline: .now() + seconds) {
-                    isFeedbackShowing = false
-                }
-                
-            }, label: {
-                HStack {
-                    Image(systemName: "checkmark.circle")
-                    Text(NSLocalizedString("confirm", comment: ""))
-                }
-            })
-        }, label: {
-            HStack {
-                Text("")
-                    .frame(width: 36, height: 36)
-                    .padding(.leading)
-                Spacer()
-                Text(NSLocalizedString("confirm", comment: ""))
-                Spacer()
-                if selectedImage != nil {
-                    Image(uiImage: selectedImage!)
-                        .resizable()
-                        .aspectRatio(contentMode: .fill)
-                        .frame(width: 36, height: 36)
-                        .cornerRadius(4)
-                        .clipped()
-                        .padding(.trailing)
-                } else {
-                    Text("")
-                        .frame(width: 36, height: 36)
-                        .padding(.trailing)
-                }
-                
-            }
-            .frame(maxWidth: .infinity, minHeight: 50)
-            .font(.system(size: 17, weight: .semibold, design: .default))
-            .foregroundColor(.white)
-            .background(Color(.link))
-        })
-        .background(Color(UIColor.systemGray6))
-        .cornerRadius(10.0)
-        .padding(.horizontal)
-        .padding(.bottom, UIScreen.main.bounds.height * 0.05)
-    }
-}
-
-struct ThemeHeaderView: View {
-    let theme: Theme
-    
-    var body: some View {
-        ZStack {
-            Rectangle()
-                .foregroundColor(.gray)
-                .frame(width: UIScreen.main.bounds.width, height: UIScreen.main.bounds.height * 0.5)
-            
-            NavigationLink(destination: ImageView(image: UIImage(named: theme.inspiration.image)!)) {
-                Image(theme.inspiration.image)
-                    .resizable()
-                    .aspectRatio(contentMode: .fill)
-                    .frame(width: UIScreen.main.bounds.width, height: UIScreen.main.bounds.height * 0.5)
-                    .clipped()
-                    .overlay(
-                        Rectangle()
-                            .foregroundColor(.clear)
-                            .background(
-                                LinearGradient(gradient: Gradient(colors: [Color.black.opacity(0.5), Color.clear, Color.clear, Color.black.opacity(0.6)]), startPoint: .top, endPoint: .bottom)
-                            )
-                    )
-            }
-            
-            VStack(alignment: .leading) {
-                Spacer()
-                HStack(alignment: .bottom) {
-                    VStack(alignment: .leading) {
-                        Text(NSLocalizedString("daily_theme", comment: ""))
-                            .textCase(.uppercase)
-                            .font(.system(size: 12, weight: .bold, design: .default))
-                        Text(theme.name)
-                            .font(.system(size: 28, weight: .bold, design: .default))
-                    }
-                    .padding(.bottom, 12)
-                    
-                    Spacer()
-                    
-                    VStack(alignment: .trailing) {
-                        Text("\(theme.inspiration.name) | \(theme.inspiration.year)")
-                            .font(.system(size: 13, weight: .bold, design: .default))
-                        
-                        Text(theme.inspiration.author)
-                            .font(.system(size: 13, weight: .regular, design: .default))
-                    }
-                    .padding(.bottom)
-                }
-                .padding(.horizontal)
-            }
-            .foregroundColor(.white)
-        }
-    }
-}
-
-struct ThemeTextView: View {
-    let theme: Theme
-    let date: Date?
-    
-    init(theme: Theme, date: Date? = nil) {
-        self.theme = theme
-        self.date = date
-    }
-    
-    var body: some View {
-        VStack(alignment: .leading) {
-            HStack {
-                if date == nil {
-                    Text(NSLocalizedString("activity_of_today", comment: ""))
-                        .font(.system(size: 20, weight: .bold, design: .default))
-                        .padding(.top, 10)
-                } else {
-                    Text(NSLocalizedString("activity_from", comment: "") + " " +
-                            DateUtils.formatToLong(date: date!, languageCode: Locale.current.languageCode == "pt" ? "pt" : "en"))
-                        .font(.system(size: 20, weight: .bold, design: .default))
-                        .padding(.top, 10)
-                }
-                Spacer()
-                
-            }
-            
-            Text(theme.description)
-                .font(.system(size: 17))
-                .padding(.top, 1)
-                .foregroundColor(.gray)
-            
-            Text(NSLocalizedString("benefits", comment: ""))
-                .font(.system(size: 20, weight: .bold, design: .default))
-                .padding(.top, 10)
-                .padding(.bottom, 5)
-            
-            if let benefits = theme.benefits {
-                ForEach(benefits, id: \.self) { benefit in
-                    Text("• \(benefit)")
-                        .padding(.bottom, 1)
-                        .foregroundColor(.gray)
-                }
-            }
-            Text("")
-                .frame(height: 100)
-            
-        }.padding()
     }
 }
